Limit options on OSS Admin/Performance page
