+++ enterprise/frontend/src/metabase-enterprise/caching/components/StrategyFormLauncher.tsx
@@ -1,11 +1,9 @@
 import { t } from "ttag";
 import { findWhere } from "underscore";
 
-import type { UpdateTargetId } from "metabase/admin/performance/strategies";
-import {
-  getShortStrategyLabel,
-  rootId,
-} from "metabase/admin/performance/strategies";
+import { rootId } from "metabase/admin/performance/constants";
+import type { UpdateTargetId } from "metabase/admin/performance/types";
+import { getShortStrategyLabel } from "metabase/admin/performance/utils";
 import { color } from "metabase/lib/colors";
 import { FixedSizeIcon, Flex, Title, Tooltip, useHover } from "metabase/ui";
 import type { Config } from "metabase-types/api";
