+++ enterprise/frontend/src/metabase-enterprise/caching/components/StrategyFormLauncherPanel.tsx
@@ -2,8 +2,8 @@ import { useMemo, type Dispatch, type SetStateAction } from "react";
 import { t } from "ttag";
 
 import { Panel } from "metabase/admin/performance/components/StrategyEditorForDatabases.styled";
-import type { UpdateTargetId } from "metabase/admin/performance/strategies";
-import { rootId } from "metabase/admin/performance/strategies";
+import { rootId } from "metabase/admin/performance/constants";
+import type { UpdateTargetId } from "metabase/admin/performance/types";
 import { FormProvider } from "metabase/forms";
 import { color } from "metabase/lib/colors";
 import { Box, Stack } from "metabase/ui";
