+++ enterprise/frontend/src/metabase-enterprise/caching/index.tsx
@@ -9,6 +9,7 @@ import QuestionCacheTTLField from "./components/QuestionCacheTTLField";
 import { SidebarCacheForm } from "./components/SidebarCacheForm";
 import { SidebarCacheSection } from "./components/SidebarCacheSection";
 import { StrategyFormLauncherPanel } from "./components/StrategyFormLauncherPanel";
+import { extraStrategies } from "./strategies";
 import {
   getQuestionsImplicitCacheTTL,
   hasQuestionCacheSection,
@@ -36,4 +37,5 @@ if (hasPremiumFeature("cache_granular_controls")) {
   PLUGIN_CACHING.DashboardStrategySidebar = DashboardStrategySidebar;
   PLUGIN_CACHING.SidebarCacheSection = SidebarCacheSection;
   PLUGIN_CACHING.SidebarCacheForm = SidebarCacheForm;
+  Object.assign(PLUGIN_CACHING.strategies, extraStrategies);
 }
