+++ frontend/src/metabase/admin/performance/constants.ts
@@ -1,0 +1,42 @@
+import * as Yup from "yup";
+
+import { PLUGIN_CACHING } from "metabase/plugins";
+
+import { isValidStrategyName } from "./utils";
+export const rootId = 0;
+
+const { strategies } = PLUGIN_CACHING;
+
+export const strategyValidationSchema = Yup.object().test(
+  "strategy-validation",
+  "The object must match one of the strategy validation schemas",
+  function (value) {
+    if (!value) {
+      return this.createError({
+        message: `Strategy is falsy`,
+      });
+    }
+    const { type } = value as unknown as { type: string };
+    if (!isValidStrategyName(type)) {
+      return this.createError({
+        message: `"${type}" is not a valid strategy name`,
+        path: "type",
+      });
+    }
+    const schema = strategies[type].validateWith;
+    try {
+      schema.validateSync(value);
+      return true;
+    } catch (error: unknown) {
+      if (error instanceof Yup.ValidationError) {
+        return this.createError({
+          message: error.message,
+          path: error.path,
+        });
+      } else {
+        console.error("Unhandled error:", error);
+        return false;
+      }
+    }
+  },
+) as Yup.AnySchema;
